+++ src/entities/schedule/ui/index.ts
@@ -1,2 +1,2 @@
-export { default as EditSchduleDropDown } from './edit-schedule-dropdown'
+export { default as EditScheduleDropDown } from './edit-schedule-dropdown'
 export { default as DeleteConfirmModal } from './delete-schedule-confirm-modal'
