+++ src/features/detailed-schedule/ui/friend-detailed-schedule.tsx
@@ -1,6 +1,6 @@
 import { useParams } from 'react-router'
 
-import { DetailedScheduleCard } from '@/entities/detailed-schedule/ui'
+import { DetailedScheduleCard, NonSchedule } from '@/entities/detailed-schedule/ui'
 import { useDetailedSchedule } from '@/entities/schedule/hooks'
 import { getRepeatText } from '@/entities/utils'
 
@@ -25,6 +25,6 @@ export default function FriendDetailedSchedule() {
       content={data.content}
     />
   ) : (
-    <div>아직 일정이 없어요</div>
+    <NonSchedule />
   )
 }
